Fix question sorting and add edit safeguards. - Remove "ordering" from questions, they're random now.

+++ src/v1/quiz/routes.ts
@@ -167,9 +167,13 @@ export default async function defineRoutes(
 
         const id = Number.parseInt(answerId);
 
+        if (id === Number.NaN) {
+          throw ApiError.invalidRequest(`${answerId} is not a valid answer ID.`);
+        }
+
         await quizdb.deleteAnswer(questionId, id);
 
-        res.status(Status.CREATED).send({ message: "Added." });
+        res.status(Status.OK).send({ message: "Deleted." });
       })
     );
